backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyPDF2==3.0.1
openai==1.35.0
httpx==0.23.0
requests==2.31.0
numpy==1.24.3
